fix: 型、ファイル構造
